fix: Redundant operations on empty collections, maps or arrays.

+++ src/main/java/gtPlusPlus/xmod/gregtech/api/metatileentity/implementations/base/generators/GregtechMetaSolarGenerator.java
@@ -188,7 +188,7 @@ public abstract class GregtechMetaSolarGenerator extends GT_MetaTileEntity_Basic
                                     && (aBaseMetaTileEntity.getUniversalEnergyStored()
                                             < ((this.maxEUOutput() * 20) + this.getMinimumStoredEU())))) {
                         this.getBaseMetaTileEntity()
-                                .increaseStoredEnergyUnits((sEnergyPerTick * this.getEfficiency()) / 10, false);
+                                .increaseStoredEnergyUnits(((long) sEnergyPerTick * this.getEfficiency()) / 10, false);
                     }
                 }
             }
